+++ app/components/index.ts
@@ -3,7 +3,7 @@ export { default as PhotoCard } from './PhotoCard';
 export { default as UploadModal } from './UploadModal';
 export { default as AudioRecorder } from './AudioRecorder';
 export { default as Camera } from './Camera';
-export { default as SpotifySearch } from './SpotifySearch';
+export { default as AppleMusicSearch } from './AppleMusicSearch';
 export { default as AuthModal } from './AuthModal';
 export { default as ProfileModal } from './ProfileModal';
 export { default as SingAlongRecorder } from './SingAlongRecorder';